Restructure datamodel
