+++ mula/scheduler/alembic/versions/0008_restructure.py
@@ -1,8 +1,8 @@
-"""Restructure
+"""restructure
 
 Revision ID: 0008
 Revises: 0007
-Create Date: 2024-03-20 11:04:50.852455
+Create Date: 2024-03-27 16:03:29.070235
 
 """
 
@@ -22,6 +22,19 @@ depends_on = None
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table(
+        "schemas",
+        sa.Column("id", scheduler.utils.datastore.GUID(), nullable=False),
+        sa.Column("scheduler_id", sa.String(), nullable=False),
+        sa.Column("hash", sa.String(length=32), nullable=True),
+        sa.Column("data", postgresql.JSONB(astext_type=sa.Text()), nullable=False),
+        sa.Column("enabled", sa.Boolean(), nullable=False),
+        sa.Column("schedule", sa.String(), nullable=True),
+        sa.Column("deadline_at", sa.DateTime(timezone=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
+        sa.Column("modified_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
         "p_items",
         sa.Column("id", scheduler.utils.datastore.GUID(), nullable=False),
         sa.Column("scheduler_id", sa.String(), nullable=True),
@@ -37,74 +50,32 @@ def upgrade():
         sa.PrimaryKeyConstraint("id"),
     )
     op.create_index(op.f("ix_p_items_hash"), "p_items", ["hash"], unique=False)
-    op.create_table(
-        "task_runs",
-        sa.Column("id", scheduler.utils.datastore.GUID(), nullable=False),
-        sa.Column("task_id", scheduler.utils.datastore.GUID(), nullable=False),
-        sa.Column(
-            "status",
-            postgresql.ENUM(
-                "PENDING",
-                "QUEUED",
-                "DISPATCHED",
-                "RUNNING",
-                "COMPLETED",
-                "FAILED",
-                "CANCELLED",
-                name="taskstatus",
-                create_type=False,
-            ),
-            nullable=False,
-        ),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
-        sa.Column("modified_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=False),
-        sa.ForeignKeyConstraint(["task_id"], ["tasks.id"], ondelete="SET NULL"),
-        sa.PrimaryKeyConstraint("id"),
-    )
     op.drop_index("ix_items_hash", table_name="items")
     op.drop_table("items")
-    op.add_column("tasks", sa.Column("hash", sa.String(length=32), nullable=True))
-    op.add_column("tasks", sa.Column("data", postgresql.JSONB(astext_type=sa.Text()), nullable=False))
-    op.add_column("tasks", sa.Column("enabled", sa.Boolean(), nullable=False))
-    op.add_column("tasks", sa.Column("schedule", sa.String(), nullable=True))
-    op.add_column("tasks", sa.Column("deadline_at", sa.DateTime(timezone=True), nullable=True))
-    op.alter_column("tasks", "scheduler_id", existing_type=sa.VARCHAR(), nullable=False)
+    op.add_column("tasks", sa.Column("schema_id", scheduler.utils.datastore.GUID(), nullable=False))
     op.drop_index("ix_tasks_p_item_hash", table_name="tasks")
-    op.drop_column("tasks", "type")
-    op.drop_column("tasks", "status")
+    op.create_foreign_key(None, "tasks", "schemas", ["schema_id"], ["id"], ondelete="SET NULL")
     op.drop_column("tasks", "p_item")
+    op.drop_column("tasks", "type")
+    op.drop_column("tasks", "scheduler_id")
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column("tasks", sa.Column("scheduler_id", sa.VARCHAR(), autoincrement=False, nullable=True))
+    op.add_column("tasks", sa.Column("type", sa.VARCHAR(), autoincrement=False, nullable=True))
     op.add_column(
         "tasks", sa.Column("p_item", postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=False)
     )
-    op.add_column(
-        "tasks",
-        sa.Column(
-            "status",
-            postgresql.ENUM(
-                "PENDING", "QUEUED", "DISPATCHED", "RUNNING", "COMPLETED", "FAILED", "CANCELLED", name="taskstatus"
-            ),
-            autoincrement=False,
-            nullable=False,
-        ),
-    )
-    op.add_column("tasks", sa.Column("type", sa.VARCHAR(), autoincrement=False, nullable=True))
+    op.drop_constraint(None, "tasks", type_="foreignkey")
     op.create_index(
         "ix_tasks_p_item_hash",
         "tasks",
         [sa.text("(p_item ->> 'hash'::text)"), sa.text("created_at DESC")],
         unique=False,
     )
-    op.alter_column("tasks", "scheduler_id", existing_type=sa.VARCHAR(), nullable=True)
-    op.drop_column("tasks", "deadline_at")
-    op.drop_column("tasks", "schedule")
-    op.drop_column("tasks", "enabled")
-    op.drop_column("tasks", "data")
-    op.drop_column("tasks", "hash")
+    op.drop_column("tasks", "schema_id")
     op.create_table(
         "items",
         sa.Column("id", sa.UUID(), autoincrement=False, nullable=False),
@@ -129,7 +100,7 @@ def downgrade():
         sa.PrimaryKeyConstraint("id", name="items_pkey"),
     )
     op.create_index("ix_items_hash", "items", ["hash"], unique=False)
-    op.drop_table("task_runs")
     op.drop_index(op.f("ix_p_items_hash"), table_name="p_items")
     op.drop_table("p_items")
+    op.drop_table("schemas")
     # ### end Alembic commands ###
