+++ mula/scheduler/storage/__init__.py
@@ -1,5 +1,5 @@
 from .filters import apply_filter
 from .pq_store import PriorityQueueStore
-from .run_store import RunStore
+from .schema_store import SchemaStore
 from .storage import DBConn, retry
 from .task_store import TaskStore
