+++ mula/tests/integration/test_schema_store.py
@@ -1,0 +1,36 @@
+import unittest
+from datetime import datetime, timedelta, timezone
+from types import SimpleNamespace
+from unittest import mock
+
+from scheduler import config, models, storage
+from scheduler.storage import filters
+from tests.factories import OrganisationFactory
+from tests.utils import functions
+
+
+class TaskStoreTestCase(unittest.TestCase):
+    def setUp(self):
+        # Application Context
+        self.mock_ctx = mock.patch("scheduler.context.AppContext").start()
+        self.mock_ctx.config = config.settings.Settings()
+
+        # Database
+        self.dbconn = storage.DBConn(str(self.mock_ctx.config.db_uri))
+        models.Base.metadata.drop_all(self.dbconn.engine)
+        models.Base.metadata.create_all(self.dbconn.engine)
+
+        self.mock_ctx.datastores = SimpleNamespace(
+            **{
+                storage.TaskStore.name: storage.TaskStore(self.dbconn),
+            }
+        )
+
+    def tearDown(self):
+        models.Base.metadata.drop_all(self.dbconn.engine)
+        self.dbconn.engine.dispose()
+
+    def test_create_task(self):
+        task = functions.create_task()
+        created_task = self.mock_ctx.datastores.task_store.create_task(task)
+        self.assertIsNotNone(created_task)
